ML_Class/requirements.txt
Pillow
torch
matplotlib
torchvision
scipy
numpy